requirements.txt
Flask
gunicorn
google-cloud-firestore
Flask-Cors